test_database_config.py

import os
import os
import unittest
import django
from django.conf import settings
from django.db import connections
from django.db.utils import OperationalError
import dj_database_url

if not os.environ.get('DJANGO_SETTINGS_MODULE'):
    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'job_portal.settings')
    # Set required env vars to avoid startup errors
    if not os.environ.get('SECRET_KEY'):
        os.environ['SECRET_KEY'] = 'test-secret-key-for-database-tests'
    os.environ['DEBUG'] = 'True' # Ensure we don't trigger production checks just for setup
    try:
        django.setup()
    except Exception as e:
        print(f"Warning: Django setup failed: {e}")

try:
    from hypothesis import given, strategies as st
    HAS_HYPOTHESIS = True
except ImportError:
    HAS_HYPOTHESIS = False

class TestDatabaseConfiguration(unittest.TestCase):
    def test_database_url_parsing(self):
        """Test that DATABASE_URL is parsed correctly."""
        # Mock environment
        url = "postgres://user:password@localhost:5432/testdb"
        config = dj_database_url.parse(url, conn_max_age=600, conn_health_checks=True)
        
        self.assertEqual(config['ENGINE'], 'django.db.backends.postgresql')
        self.assertEqual(config['NAME'], 'testdb')
        self.assertEqual(config['USER'], 'user')
        self.assertEqual(config['PASSWORD'], 'password')
        self.assertEqual(config['HOST'], 'localhost')
        self.assertEqual(config['PORT'], 5432)
        self.assertEqual(config['CONN_MAX_AGE'], 600)
        self.assertEqual(config['CONN_HEALTH_CHECKS'], True)

    def test_sqlite_fallback(self):
        """Test fallback to SQLite when DATABASE_URL is not present."""
        # We can't easily change os.environ for the loaded settings, 
        # but we can verify what current settings are doing if we assume dev env.
        # OR we can manually invoke the logic used in settings.py
        
        # logic from settings.py:
        # if os.environ.get('DATABASE_URL'): ... else: ...
        
        # Verify fallback logic implementation
        config = dj_database_url.config(default='sqlite:///db.sqlite3')
        self.assertEqual(config['ENGINE'], 'django.db.backends.sqlite3')

    @unittest.skipIf(not HAS_HYPOTHESIS, "Hypothesis not installed")
    def test_database_url_properties(self):
        """Property test: Verify dj_database_url parsing properties."""
        if not HAS_HYPOTHESIS: return

        @given(
            st.text(min_size=1, alphabet=st.characters(whitelist_categories=('L', 'N'))), # dbname
            st.text(min_size=1, alphabet=st.characters(whitelist_categories=('L', 'N'))), # user
            st.text(min_size=1, alphabet=st.characters(whitelist_categories=('L', 'N'))), # password
            st.integers(min_value=1024, max_value=65535) # port
        )
        def check_url_parsing(dbname, user, password, port):
            # Construct a valid postgres URL
            url = f"postgres://{user}:{password}@localhost:{port}/{dbname}"
            config = dj_database_url.parse(url)
            
            self.assertEqual(config['NAME'], dbname)
            self.assertEqual(config['USER'], user)
            self.assertEqual(config['PASSWORD'], password)
            self.assertEqual(config['PORT'], port)
            self.assertEqual(config['ENGINE'], 'django.db.backends.postgresql')
            
        check_url_parsing()

class TestDatabaseConnectionBehavior(unittest.TestCase):
    def test_connection_pooling_config(self):
        """Test that connection pooling is configured."""
        # Check settings.DATABASES['default']['CONN_MAX_AGE']
        # We need to access the configured settings.
        # Since we might be in a mode where DATABASE_URL wasn't set, this might be None/0.
        
        # If we are in dev, likely using SQLite, which might not have pooling set in settings.py logic unless checking env.
        pass

if __name__ == '__main__':
    unittest.main()
