requirements-production.txt
# Core Django framework
Django==6.0.1
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.2
pytz==2025.2

# Database support
dj-database-url==3.1.0
psycopg2-binary==2.9.11

# Authentication and user management
django-allauth==65.14.0

# Forms and UI
django-crispy-forms==2.5
crispy-bootstrap5==2025.6

# Static file serving
whitenoise==6.11.0

# File handling
pillow==12.0.0
pdfminer.six==20260107

# Date/time utilities
python-dateutil==2.9.0.post0

# Production server
gunicorn==25.0.0